fix(lite): fix input invalid bug in lar for fitting mode
GitOrigin-RevId: 45d81c9a96dc64c4f41dd87581a5fa2bca76c2b7

+++ lite/load_and_run/examples/script/mge_input_data.py
@@ -26,7 +26,7 @@ cv2.imwrite("input_data/cat.ppm",processed_img)
 #json 
 data_obj = {
     "shape": [1,3],
-    "type": "int32",
+    "type": "float32",
     "raw": [2,3,4]
 }
 with open("input_data/add_demo_input.json", "w") as f:
